Move HeapkeyValueStore to model mod
