+++ radix-engine/src/engine/mod.rs
@@ -8,7 +8,6 @@ mod native_interpreter;
 mod node;
 mod node_properties;
 mod node_ref;
-mod precommitted_kv_store;
 mod system_api;
 mod track;
 mod track_support;
@@ -24,7 +23,6 @@ pub use native_interpreter::NativeInterpreter;
 pub use node::*;
 pub use node_properties::*;
 pub use node_ref::*;
-pub use precommitted_kv_store::*;
 pub use system_api::SystemApi;
 pub use track::*;
 pub use track_support::*;
