+++ radix-engine/src/model/mod.rs
@@ -6,6 +6,7 @@ mod component;
 mod method_authorization;
 mod non_fungible;
 mod package_extractor;
+mod precommitted_kv_store;
 mod proof;
 mod resource;
 mod resource_manager;
@@ -14,6 +15,7 @@ mod transaction_processor;
 mod validated_package;
 mod vault;
 mod worktop;
+mod wrappers;
 
 pub use abi_extractor::*;
 pub use auth_converter::convert;
@@ -26,6 +28,7 @@ pub use method_authorization::{
 };
 pub use non_fungible::NonFungible;
 pub use package_extractor::{extract_package, ExtractAbiError};
+pub use precommitted_kv_store::HeapKeyValueStore;
 pub use proof::*;
 pub use resource::*;
 pub use resource_manager::{ResourceManager, ResourceManagerError};
@@ -36,16 +39,4 @@ pub use transaction_processor::{
 pub use validated_package::{PackageError, ValidatedPackage};
 pub use vault::{Vault, VaultError};
 pub use worktop::{Worktop, WorktopError};
-
-use sbor::rust::vec::Vec;
-use sbor::*;
-
-/// To support non-fungible deletion, we wrap it into a container
-/// when persisting into the substate store.
-#[derive(Debug, Clone, TypeId, Encode, Decode, PartialEq, Eq)]
-pub struct NonFungibleWrapper(pub Option<NonFungible>);
-
-/// To support key value store entry deletion, we wrap it into a container
-/// when persisting into the substate store.
-#[derive(Debug, Clone, TypeId, Encode, Decode, PartialEq, Eq)]
-pub struct KeyValueStoreEntryWrapper(pub Option<Vec<u8>>);
+pub use wrappers::*;
