+++ radix-engine/src/model/wrappers.rs
@@ -1,0 +1,14 @@
+use sbor::rust::vec::Vec;
+use sbor::*;
+
+use crate::model::NonFungible;
+
+/// To support non-fungible deletion, we wrap it into a container
+/// when persisting into the substate store.
+#[derive(Debug, Clone, TypeId, Encode, Decode, PartialEq, Eq)]
+pub struct NonFungibleWrapper(pub Option<NonFungible>);
+
+/// To support key value store entry deletion, we wrap it into a container
+/// when persisting into the substate store.
+#[derive(Debug, Clone, TypeId, Encode, Decode, PartialEq, Eq)]
+pub struct KeyValueStoreEntryWrapper(pub Option<Vec<u8>>);
